Create new/edit existing recipe functionality
Added the ability to edit and create recipes. I can't get the rest of the site to load on my machine, so it hasn't been tested properly in its little frame home.
Feed the page a recipeID of -1 and you'll be presented with a new recipe, with the edit functionality turned on.